Update TrafficCop for wnp 120
